组图新闻的布局优化

+++ app/src/main/java/com/rayho/tsxiu/module_news/bean/NewsBean.java
@@ -54,7 +54,7 @@ public class NewsBean {
         public String catPathKey;
         public String title;
         public String id;
-        public int publishDate;
+        public long publishDate;
         public int dislikeCount;
         public int commentCount;
         public int likeCount;
